pyproject.toml
[tool.poetry]
package-mode = false
name = "cantus-ultimus"

[tool.poetry.dependencies]
python = "^3.8"
celery = "5.3.1"
django = "4.2.11"
django-extensions = "3.2.3"
django-celery-results = "2.5.1"
djangorestframework = "3.14.0"
gunicorn = "20.1.0"
lxml = "5.1.0"
markdown = "3.4.3"
psycopg = "3.1.8"
requests = "2.31.0"
solrpy = "0.9.9"
Werkzeug = "3.0.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^24.1.0"
mypy = "^1.9.0"
pylint = "^3.1.0"
lxml-stubs = "^0.5.1"
django-stubs = "^4.2.7"
pylint-django = "^2.5.5"
djangorestframework-stubs = "^3.14.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
